Big Refactoring
Split Actionbar when narrow
Remove the ToggleButton to start the DoS, instead use
the action bar to toggle the Stress-Test on and off.

Move InformationFragment to DosFragment, but dont remove
InformationFragment for later use.

Fix Power of the Stress-Test
due the url in threadinject doesnt contain "http://" it
cant establish a connection and thus nothing will happen

Change-Id: I8005bf83be74202f2078daa0c4359344d51170b6